+++ api/builtin/welcome.ts
