+++ api/controllers/index.ts
@@ -16,7 +16,7 @@ import * as actions from './actions'
 import {checkTag} from '../utils/gitinfo'
 import * as path from 'path'
 import * as timers from '../utils/timers'
-import * as builtInBots from '../bots'
+import * as builtInBots from '../builtin'
 
 const env = process.env.NODE_ENV || 'development';
 console.log("=> env:",env)
